[1215172109860419] refactor: split Home.tsx monolith into focused modules
Break the 2,438-line Home.tsx into a Home/ directory with 8 focused files:
- types.ts (161 lines): All shared type definitions
- constants.ts (9 lines): MOBILE_CONFIG, OPEN_THREAD_KEY
- HomeFeedEmbed.tsx (674 lines): Embed rendering (images, video, quotes, etc.)
- PostItem.tsx (316 lines): Individual post card rendering
- useFeedQuery.ts (315 lines): Feed fetching with offline/cache support
- useHomeKeyboard.ts (290 lines): Keyboard navigation + shortcut registration
- useFeedOptions.ts (150 lines): Feed preferences and options building
- index.tsx (537 lines): Slim orchestration component

The main component dropped from 2,438 to 537 lines (78% reduction).
Public API unchanged — SkyColumn.tsx import resolves via index.tsx.
